\name{physical}
\alias{physical}
\docType{data}
\title{
Physical strength of males
}
\description{
Physical strength of males
}
\usage{data(physical)}
\format{
4592 contests between 82 stimuli (average of 112 appearances per stimulus). 56 raters came to the lab and made 82 judgements each. They were asked to choose which image of a pair of stimulus images depicted the physically stronger looking male.
  \describe{
    \item{\code{Date}}{Date of the rating}
    \item{\code{Winner}}{Winner of the interaction}
    \item{\code{Loser}}{Loser of the interaction}
  }
}
\details{
4592 contests between 82 stimuli (average of 112 appearances per stimulus). 56 raters came to the lab and made 82 judgements each. They were asked to choose which image of a pair of stimulus images depicted the physically stronger looking male.
}
\source{
Andrew Clark
}
\references{
Andrew Clark
}
\examples{
data(physical)
}
\keyword{datasets}
