% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.elochoice.R
\name{summary.elochoice}
\alias{summary.elochoice}
\title{summarize elochoice object}
\usage{
\method{summary}{elochoice}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"elochoice"}, usually the result of a call to \code{\link{elochoice}}}

\item{...}{further arguments passed to or from other methods. Nothing relevant in this case.}
}
\description{
summarize elochoice object
}
\examples{
xdata <- randompairs(nstim=10, nint=500)
x <- elochoice(xdata$winner, xdata$loser, runs=5)
summary(x)
}
\author{
Christof Neumann
}
