% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pysical.R
\docType{data}
\name{physical}
\alias{physical}
\title{Physical strength of males}
\format{4592 pairwise comparisons (contests) between 82 stimuli (average of 112 appearances per stimulus). 56 raters came to the lab and made 82 judgements each. They were asked to choose which image of a pair of stimulus images depicted the physically stronger looking male.
  \describe{
    \item{\code{Date}}{Date of the rating}
    \item{\code{Winner}}{Winner of the interaction}
    \item{\code{Loser}}{Loser of the interaction}
    \item{\code{raterID}}{A numeric indicator of rater identity}
   }}
\source{
Andrew Clark
}
\usage{
data(physical)
}
\description{
Physical strength of males
}
\examples{
data(physical)
}
\references{
Andrew Clark
}
\keyword{datasets}
